Moves playrho::Wider to playrho::d2::Wider & adds playrho::WiderType alias
